refactor(Metrics): :recycle: Change approach to query requests
